refactor(api): Migrate API functions from apiManager to dedicated modules for authentication, user, connections, and workshop. Update imports across components to reflect new API paths. Remove deprecated apiManager and related files. Clean up unused styles and optimize code formatting in various components.
