+++ src/components/me/accounts/Minecraft.tsx
@@ -7,19 +7,18 @@ import {
     IconX
 } from '@tabler/icons-react';
 import Image from 'next/image';
-import {
-    connectMinecraft as connectMinecraftAPI,
-    disconnectMinecraft,
-    getMeConnections,
-    purgeSkinCache,
-    setMinecraftAutoload,
-    setMinecraftVisible
-} from '@/lib/apiManager';
 import { minecraftMono } from '@/fonts/Minecraft';
 import SlideButton from '@/components/SlideButton';
 import { formatDateHuman } from '@/lib/time';
 import MinecraftConnect from '../MinecraftConnect';
 import DisconnectHelper from './DisconnectHelper';
+import {
+    connectMinecraft,
+    disconnectMinecraft,
+    getMeConnections
+} from '@/lib/api/connections';
+import { setUserSetting } from '@/lib/api/user';
+import { purgeSkinCache } from '@/lib/api/minecraft';
 
 export const Minecraft = () => {
     const { data, isLoading, mutate } = useSWR(
@@ -57,7 +56,7 @@ export const Minecraft = () => {
                     label="Отображать ник в поиске"
                     defaultValue={data.minecraft.valid}
                     onChange={async state => {
-                        await setMinecraftVisible({ state });
+                        await setUserSetting({ nick_search: state });
                         mutate({
                             ...data,
                             minecraft: {
@@ -73,7 +72,7 @@ export const Minecraft = () => {
                     label="Автоматически устанавливать скин в редакторе"
                     defaultValue={data.minecraft.autoload}
                     onChange={async state => {
-                        await setMinecraftAutoload({ state });
+                        await setUserSetting({ skin_autoload: state });
                         mutate({
                             ...data,
                             minecraft: {
@@ -120,7 +119,7 @@ const UnloggedMinecraft = () => {
                 Подключите свой аккаунт Minecraft, чтобы управлять кэшем скинов и
                 настройками видимости вашего никнейма в поиске.
             </p>
-            <MinecraftConnect onInput={connectMinecraft}>
+            <MinecraftConnect onInput={connect}>
                 <button className={Style.unlink} style={{ width: '100%' }}>
                     <IconPlugConnected />
                     Подключить
@@ -151,7 +150,7 @@ const disconnect = () => {
         .catch(console.error);
 };
 
-const connectMinecraft = async (code: string): Promise<void> => {
-    await connectMinecraftAPI(code);
+const connect = async (code: string): Promise<void> => {
+    await connectMinecraft(code);
     setTimeout(() => mutate('userConnections'), 150);
 };
