+++ src/lib/api/auth.ts
@@ -1,0 +1,39 @@
+import { doRequest } from './utils';
+
+export const loginDiscord = async (code: string) => {
+    return (
+        await doRequest({
+            url: `/auth/discord/${code}`,
+            method: 'POST'
+        })
+    ).data;
+};
+
+export const loginMinecraft = async (code: string) => {
+    return (
+        await doRequest({
+            url: `/auth/minecraft/${code}`,
+            method: 'POST'
+        })
+    ).data;
+};
+
+export const loginGoogle = async (code: string) => {
+    return (
+        await doRequest({
+            url: `/auth/google`,
+            method: 'POST',
+            data: { code }
+        })
+    ).data;
+};
+
+export const loginTwitch = async (code: string) => {
+    return (
+        await doRequest({
+            url: `/auth/twitch`,
+            method: 'POST',
+            data: { code }
+        })
+    ).data;
+};
