--- src/styles/RolesDialog.module.css
@@ -1,29 +1,0 @@
-.roles_container {
-    display: flex;
-    flex-wrap: wrap;
-}
-
-.link {
-    display: flex;
-    justify-content: center;
-    border: 2px solid var(--hr-color);
-    border-radius: 10px;
-    height: 2.3rem;
-    align-items: center;
-    box-sizing: border-box;
-    text-decoration: none;
-    transition-timing-function: cubic-bezier(.4, 0, .2, 1);
-    transition-duration: .15s;
-    gap: .2rem;
-}
-
-.link:hover {
-    background-color: var(--accent);
-}
-
-.container {
-    display: flex;
-    flex-direction: column;
-    align-items: stretch;
-    gap: .4rem;
-}
