+++ src/styles/admin/page.module.css
@@ -30,12 +30,6 @@
     gap: .5rem
 }
 
-.register_container {
-    display: flex;
-    gap: .5rem;
-    flex-wrap: wrap
-}
-
 .users_container {
     display: flex;
     flex-direction: column;
