+++ src/styles/root/page.module.css
@@ -124,31 +124,6 @@
     border: 2px rgba(12, 247, 215) dashed;
 }
 
-.migrationOpenButton {
-    font-family: inherit;
-    cursor: pointer;
-    animation: migrationFading 2s infinite;
-    font-weight: 500;
-}
-
-.migrationOpenButton:hover {
-    border: 2px #D29922 dashed !important;
-}
-
-@keyframes migrationFading {
-    0% {
-        border-color: var(--main-element-color);
-    }
-
-    50% {
-        border-color: #D29922;
-    }
-
-    100% {
-        border-color: var(--main-element-color);
-    }
-}
-
 .footer {
     position: fixed;
     bottom: 1rem;
